Rework miner test setups to fix tests

+++ cmd/lotus-seed/main.go
@@ -61,7 +61,7 @@ var preSealCmd = &cli.Command{
 			Value: "lotus is fire",
 			Usage: "set the ticket preimage for sealing randomness",
 		},
-		&cli.Uint64Flag{
+		&cli.IntFlag{
 			Name:  "num-sectors",
 			Value: 1,
 			Usage: "select number of sectors to pre-seal",
@@ -79,7 +79,7 @@ var preSealCmd = &cli.Command{
 			return err
 		}
 
-		gm, err := seed.PreSeal(maddr, c.Uint64("sector-size"), c.Uint64("num-sectors"), sbroot, []byte(c.String("ticket-preimage")))
+		gm, err := seed.PreSeal(maddr, c.Uint64("sector-size"), c.Int("num-sectors"), sbroot, []byte(c.String("ticket-preimage")))
 		if err != nil {
 			return err
 		}
